feat: implement `CitiBuy.get_invoices()` method (#97)
* feat: Adds Invoice and PO status mapping to CitiBuy ETL class

* refactor: Reorganizes CitiBuy test data to split input from output data

* feat: Adds InvoiceStatusHistory to models
Creates the InvoiceStatusHistory model class which represents the 
INVOICE_STATUS_DATES table in CitiBuy and adds test data for this new 
table to citibuy_data.py

* refactor: Updates the datatype from float to numeric in CitiBuy models

* feat: Implements CitiBuy.get_invoices() method

closes: #32

+++ app/src/dgs_fiscal/systems/citibuy/models/base.py
@@ -4,7 +4,7 @@ from sqlalchemy import (
     Column,
     Integer,
     String,
-    Float,
+    Numeric,
     DateTime,
     ForeignKey,
     ForeignKeyConstraint,
